refactor(webapp): split the icon barrel into one module per icon
Every icon exported from the Figma file lived in a single figma-icons.tsx. Give each its own module, named
the way the icons supplied outside that file already were, and delete the
barrel. Repoint the rail, the account menu, the project row, the global
dashboard and the settings page at the new modules.

Declare the props contract once as FigmaIconProps, replacing the local
copy each icon carried, and reduce every icon's comment to a line saying
what the glyph shows — the sizes, insets, hex fills and node ids they
listed described the design file rather than the code, and went stale the
moment either changed.
