+++ webapp/src/components/account-menu.tsx
@@ -8,7 +8,8 @@ import { fetcher } from "@/api/swr";
 import { cn } from "@/helpers/utils";
 import { useModal } from "@/hooks/useModal";
 import CreateOrganizationModal from "./createOrganizationModal";
-import { ExitToAppIcon, SettingsIcon } from "./icons/figma-icons";
+import { ExitToAppIcon } from "./icons/exit-to-app-icon";
+import { SettingsIcon } from "./icons/settings-icon";
 import { OrganizationIcon } from "./icons/organization-icon";
 import { PlusIcon } from "./icons/plus-icon";
 import { DropdownMenu, DropdownMenuTrigger } from "./ui/dropdown-menu";
