+++ webapp/src/components/icons/account-circle-icon.tsx
@@ -1,0 +1,24 @@
+import { FigmaIconProps } from "./types";
+
+/*
+ * Account in circle icon — a head and shoulders inside a circle.
+ */
+
+export function AccountCircleIcon({ className }: FigmaIconProps) {
+    return (
+        <svg
+            viewBox="0 0 24 24"
+            width="24"
+            height="24"
+            fill="currentColor"
+            xmlns="http://www.w3.org/2000/svg"
+            aria-hidden="true"
+            focusable="false"
+            className={className}
+        >
+            <g transform="translate(3 3)">
+                <path d="M3.19625 14.4845C4.04625 13.873 4.94817 13.3894 5.902 13.0337C6.85583 12.6779 7.8885 12.5 9 12.5C10.1115 12.5 11.1442 12.6779 12.098 13.0337C13.0518 13.3894 13.9538 13.873 14.8038 14.4845C15.4641 13.8012 15.9952 12.9942 16.397 12.0635C16.799 11.1327 17 10.1115 17 9C17 6.78333 16.2208 4.89583 14.6625 3.3375C13.1042 1.77917 11.2167 1 9 1C6.78333 1 4.89583 1.77917 3.3375 3.3375C1.77917 4.89583 1 6.78333 1 9C1 10.1115 1.201 11.1327 1.603 12.0635C2.00483 12.9942 2.53592 13.8012 3.19625 14.4845ZM9.0005 9.5C8.15783 9.5 7.44708 9.21075 6.86825 8.63225C6.28942 8.05375 6 7.34317 6 6.5005C6 5.65783 6.28925 4.94708 6.86775 4.36825C7.44625 3.78942 8.15683 3.5 8.9995 3.5C9.84217 3.5 10.5529 3.78925 11.1318 4.36775C11.7106 4.94625 12 5.65683 12 6.4995C12 7.34217 11.7107 8.05292 11.1322 8.63175C10.5537 9.21058 9.84317 9.5 9.0005 9.5ZM9 18C7.74483 18 6.56983 17.7663 5.475 17.299C4.38017 16.8317 3.42758 16.1929 2.61725 15.3828C1.80708 14.5724 1.16833 13.6198 0.701 12.525C0.233667 11.4302 0 10.2552 0 9C0 7.74483 0.233667 6.56983 0.701 5.475C1.16833 4.38017 1.80708 3.42758 2.61725 2.61725C3.42758 1.80708 4.38017 1.16833 5.475 0.700999C6.56983 0.233666 7.74483 0 9 0C10.2552 0 11.4302 0.233666 12.525 0.700999C13.6198 1.16833 14.5724 1.80708 15.3828 2.61725C16.1929 3.42758 16.8317 4.38017 17.299 5.475C17.7663 6.56983 18 7.74483 18 9C18 10.2552 17.7663 11.4302 17.299 12.525C16.8317 13.6198 16.1929 14.5724 15.3828 15.3828C14.5724 16.1929 13.6198 16.8317 12.525 17.299C11.4302 17.7663 10.2552 18 9 18ZM9 17C9.92183 17 10.8289 16.8388 11.7212 16.5163C12.6134 16.1939 13.3845 15.7526 14.0345 15.1923C13.3845 14.6704 12.6326 14.2579 11.7788 13.9548C10.9249 13.6516 9.99867 13.5 9 13.5C8.00133 13.5 7.07183 13.6484 6.2115 13.9453C5.35133 14.2421 4.60267 14.6578 3.9655 15.1923C4.6155 15.7526 5.38658 16.1939 6.27875 16.5163C7.17108 16.8388 8.07817 17 9 17ZM9 8.5C9.5615 8.5 10.0352 8.30708 10.4212 7.92125C10.8071 7.53525 11 7.0615 11 6.5C11 5.9385 10.8071 5.46475 10.4212 5.07875C10.0352 4.69292 9.5615 4.5 9 4.5C8.4385 4.5 7.96475 4.69292 7.57875 5.07875C7.19292 5.46475 7 5.9385 7 6.5C7 7.0615 7.19292 7.53525 7.57875 7.92125C7.96475 8.30708 8.4385 8.5 9 8.5Z" />
+            </g>
+        </svg>
+    );
+}
