+++ webapp/src/components/icons/add-a-photo-icon.tsx
@@ -1,0 +1,22 @@
+import { FigmaIconProps } from "./types";
+
+/*
+ * Add-a-photo icon — a camera with a plus sign at its corner.
+ */
+
+export function AddAPhotoIcon({ className }: FigmaIconProps) {
+    return (
+        <svg
+            viewBox="0 0 21.5384 21.5384"
+            width="21.5384"
+            height="21.5384"
+            fill="currentColor"
+            xmlns="http://www.w3.org/2000/svg"
+            aria-hidden="true"
+            focusable="false"
+            className={className}
+        >
+            <path d="M2.69208 18.8462C2.19849 18.8462 1.77595 18.6704 1.42445 18.3189C1.07296 17.9674 0.89721 17.5449 0.89721 17.0513V6.28211C0.89721 5.78852 1.07296 5.36598 1.42445 5.01449C1.77595 4.66299 2.19849 4.48724 2.69208 4.48724H5.51899L7.17925 2.69238H12.5639V4.48724H7.9645L6.32669 6.28211H2.69208V17.0513H17.051V8.97441H18.8459V17.0513C18.8459 17.5449 18.6701 17.9674 18.3186 18.3189C17.9672 18.6704 17.5446 18.8462 17.051 18.8462H2.69208ZM17.051 6.28211V4.48724H15.2562V2.69238H17.051V0.897509H18.8459V2.69238H20.6408V4.48724H18.8459V6.28211H17.051ZM9.87155 15.7052C10.9933 15.7052 11.9469 15.3125 12.7321 14.5273C13.5174 13.742 13.91 12.7885 13.91 11.6667C13.91 10.5449 13.5174 9.5914 12.7321 8.80614C11.9469 8.02089 10.9933 7.62826 9.87155 7.62826C8.74976 7.62826 7.79623 8.02089 7.01098 8.80614C6.22572 9.5914 5.8331 10.5449 5.8331 11.6667C5.8331 12.7885 6.22572 13.742 7.01098 14.5273C7.79623 15.3125 8.74976 15.7052 9.87155 15.7052ZM9.87155 13.9103C9.24335 13.9103 8.71236 13.6934 8.2786 13.2597C7.84484 12.8259 7.62796 12.2949 7.62796 11.6667C7.62796 11.0385 7.84484 10.5075 8.2786 10.0738C8.71236 9.64001 9.24335 9.42313 9.87155 9.42313C10.4998 9.42313 11.0307 9.64001 11.4645 10.0738C11.8983 10.5075 12.1151 11.0385 12.1151 11.6667C12.1151 12.2949 11.8983 12.8259 11.4645 13.2597C11.0307 13.6934 10.4998 13.9103 9.87155 13.9103Z" />
+        </svg>
+    );
+}
