+++ webapp/src/components/icons/exit-to-app-icon.tsx
@@ -1,0 +1,24 @@
+import { FigmaIconProps } from "./types";
+
+/*
+ * Log-out icon — an arrow leaving through an open door.
+ */
+
+export function ExitToAppIcon({ className }: FigmaIconProps) {
+    return (
+        <svg
+            viewBox="0 0 24 24"
+            width="24"
+            height="24"
+            fill="currentColor"
+            xmlns="http://www.w3.org/2000/svg"
+            aria-hidden="true"
+            focusable="false"
+            className={className}
+        >
+            <g transform="translate(3.701 3.701)">
+                <path d="M1.705 16.596C1.23217 16.596 0.82975 16.4301 0.49775 16.0982C0.165917 15.7662 0 15.3632 0 14.8892V11.2787H1.399V14.8892C1.399 14.9662 1.431 15.0368 1.495 15.101C1.55917 15.165 1.62975 15.197 1.70675 15.197H14.8892C14.9662 15.197 15.0368 15.165 15.101 15.101C15.165 15.0368 15.197 14.9662 15.197 14.8892V1.70675C15.197 1.62975 15.165 1.55917 15.101 1.495C15.0368 1.431 14.9662 1.399 14.8892 1.399H1.70675C1.62975 1.399 1.55917 1.431 1.495 1.495C1.431 1.55917 1.399 1.62975 1.399 1.70675V5.3125H0V1.70675C0 1.23275 0.165917 0.829749 0.49775 0.497749C0.82975 0.165916 1.23217 0 1.705 0H14.891C15.3638 0 15.7662 0.165916 16.0982 0.497749C16.4301 0.829749 16.596 1.23217 16.596 1.705V14.891C16.596 15.3638 16.4301 15.7662 16.0982 16.0982C15.7662 16.4301 15.3638 16.596 14.891 16.596H1.705ZM6.9835 12.6382L5.998 11.6017L8.60475 8.995H0V7.596H8.60475L5.998 4.98925L6.9835 3.95775L11.3037 8.30575L6.9835 12.6382Z" />
+            </g>
+        </svg>
+    );
+}
