--- webapp/src/components/icons/figma-icons.tsx
@@ -1,284 +1,0 @@
-/*
- * Icons exported verbatim from the Code Carbon Figma file.
- *
- * Every path below is the exact "d" attribute Figma produced — nothing was
- * redrawn, simplified, or swapped for a Lucide/Heroicons equivalent. Figma
- * gives each of these a single flat fill; that fill is replaced with
- * `currentColor` so the caller controls state colour (the rail icons are
- * #BFFB4F when selected and #FFFFFF otherwise, which is a colour change on one
- * shared vector, not two different vectors).
- *
- * The Material-derived icons (menu, settings, exit_to_app, add_a_photo) ship
- * from Figma as a glyph clipped by an alpha mask whose only content is the
- * bounding-box rect. That mask is a no-op, so it is flattened away here; the
- * glyph is instead placed at its exact Figma offset inside the outer box, which
- * preserves both the outer box and the inner leaf geometry.
- */
-
-type IconProps = {
-    className?: string;
-};
-
-/** 32x32. Lime (#BFFB4F) in the selected rail item. */
-export function GlobalIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 32 32"
-            width="32"
-            height="32"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M22.85 7.62V10.67H21.33V15.24H22.85V16.76H24.38V18.29H19.81V19.81H18.28V21.33H22.85V24.38H21.33V27.43H22.85V28.95H25.9V27.43H27.43V25.9H28.95V22.86H30.47V19.81H32V12.19H30.47V9.14H28.95V6.1H27.43V7.62H22.85Z" />
-            <path d="M27.43 4.57H25.9V6.1H27.43V4.57Z" />
-            <path d="M25.9 3.05H22.85V4.57H25.9V3.05Z" />
-            <path d="M22.85 28.95H19.81V30.48H22.85V28.95Z" />
-            <path d="M22.85 1.52H19.81V3.05H22.85V1.52Z" />
-            <path d="M16.76 28.95H15.24V27.43H10.66V28.95H9.14V30.48H12.19V32H19.81V30.48H16.76V28.95Z" />
-            <path d="M19.81 0H12.19V1.52H19.81V0Z" />
-            <path d="M10.66 16.76H12.19V18.29H13.71V13.71H10.66V16.76Z" />
-            <path d="M9.14 27.43H6.09V28.95H9.14V27.43Z" />
-            <path d="M6.09 25.9H4.57V27.43H6.09V25.9Z" />
-            <path d="M3.05 22.86V25.9H4.57V24.38H6.09V22.86H7.62V18.29H6.09V16.76H4.57V15.24H3.05V13.71H4.57V12.19H9.14V10.67H10.66V9.14H12.19V10.67H13.71V9.14H15.24V6.1H13.71V4.57H10.66V3.05H12.19V1.52H9.14V3.05H6.09V4.57H4.57V6.1H3.05V9.14H1.52V12.19H0V19.81H1.52V22.86H3.05Z" />
-        </svg>
-    );
-}
-
-/** Projects — 32x32. */
-export function ProjectsIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 32 32"
-            width="32"
-            height="32"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M28.96 6.1V0H13.72V1.52H7.62V3.05H3.05V13.72H1.53V16.76H0V32H32V6.1H28.96ZM7.62 30.48H1.53V21.33H7.62V30.48ZM7.62 19.81H4.58V4.57H7.62V19.81ZM9.15 3.05H13.72V4.57H10.67V19.81H9.15V3.05ZM13.72 6.1V19.81H12.2V6.1H13.72ZM15.24 30.48H9.15V21.33H15.24V30.48ZM30.48 30.48H16.77V19.81H15.24V1.52H27.43V6.1H22.86V7.62H30.48V30.48Z" />
-            <path d="M27.43 21.33H25.91V22.86H27.43V21.33Z" />
-            <path d="M27.43 18.29H25.91V19.81H27.43V18.29Z" />
-            <path d="M25.91 22.86H22.86V24.38H25.91V22.86Z" />
-            <path d="M22.86 21.33H21.34V22.86H22.86V21.33Z" />
-            <path d="M22.86 18.29H21.34V19.81H22.86V18.29Z" />
-            <path d="M22.86 7.62H21.34V10.67H22.86V7.62Z" />
-            <path d="M21.34 10.67H19.81V13.72H21.34V10.67Z" />
-            <path d="M19.81 13.72H18.29V16.76H19.81V13.72Z" />
-            <path d="M19.81 3.05H18.29V4.57H19.81V3.05Z" />
-            <path d="M18.29 16.76H16.77V19.81H18.29V16.76Z" />
-            <path d="M13.72 24.38H10.67V27.43H13.72V24.38Z" />
-            <path d="M6.1 24.38H3.05V27.43H6.1V24.38Z" />
-        </svg>
-    );
-}
-
-/** Glossary — 32x32. */
-export function GlossaryIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 32 32"
-            width="32"
-            height="32"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M30.47 24.38H28.95V21.33H27.42V24.38H25.9V25.9H24.38V30.48H25.9V32H30.47V30.48H32V25.9H30.47V24.38Z" />
-            <path d="M27.42 19.81H25.9V21.33H27.42V19.81Z" />
-            <path d="M16.76 18.29V13.71H19.8V12.19H12.19V13.71H15.23V18.29H6.09V19.81H15.23V24.38H13.71V25.9H12.19V30.48H13.71V32H18.28V30.48H19.8V25.9H18.28V24.38H16.76V19.81H25.9V18.29H16.76Z" />
-            <path d="M22.85 3.05H21.33V10.67H22.85V3.05Z" />
-            <path d="M21.33 10.67H19.8V12.19H21.33V10.67Z" />
-            <path d="M21.33 1.52H19.8V3.05H21.33V1.52Z" />
-            <path d="M19.8 7.62H18.28V9.14H19.8V7.62Z" />
-            <path d="M19.8 4.57H18.28V6.09H19.8V4.57Z" />
-            <path d="M18.28 9.14H13.71V10.67H18.28V9.14Z" />
-            <path d="M19.8 0H12.19V1.52H19.8V0Z" />
-            <path d="M13.71 7.62H12.19V9.14H13.71V7.62Z" />
-            <path d="M13.71 4.57H12.19V6.09H13.71V4.57Z" />
-            <path d="M12.19 10.67H10.66V12.19H12.19V10.67Z" />
-            <path d="M12.19 1.52H10.66V3.05H12.19V1.52Z" />
-            <path d="M10.66 3.05H9.14V10.67H10.66V3.05Z" />
-            <path d="M6.09 19.81H4.57V21.33H6.09V19.81Z" />
-            <path d="M3.04 24.38H1.52V25.9H0V30.48H1.52V32H6.09V30.48H7.61V25.9H6.09V24.38H4.57V21.33H3.04V24.38Z" />
-        </svg>
-    );
-}
-
-/** 24x24 hamburger, fill #666666 in the design. */
-export function MenuIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M3 18V16H21V18H3ZM3 13V11H21V13H3ZM3 8V6H21V8H3Z" />
-        </svg>
-    );
-}
-
-/**
- * 24x24 outer box holding an 18.792x19 glyph, inset 10.85%
- * horizontally (2.604px) and 10.42% vertically (2.5px). Fill #949494.
- */
-export function SettingsIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <g transform="translate(2.604 2.5)">
-                <path d="M7.08825 19L6.7075 15.9538C6.43967 15.8641 6.165 15.7385 5.8835 15.577C5.60217 15.4153 5.35058 15.2422 5.12875 15.0577L2.3075 16.25L0 12.25L2.44025 10.4058C2.41725 10.2571 2.40092 10.1078 2.39125 9.95775C2.38158 9.80775 2.37675 9.65833 2.37675 9.5095C2.37675 9.36733 2.38158 9.22283 2.39125 9.076C2.40092 8.92917 2.41725 8.76858 2.44025 8.59425L0 6.75L2.3075 2.76925L5.119 3.952C5.36017 3.761 5.61758 3.58633 5.89125 3.428C6.16492 3.26967 6.43383 3.14242 6.698 3.04625L7.08825 0H11.7037L12.0845 3.05575C12.3845 3.16475 12.6559 3.292 12.8987 3.4375C13.1417 3.583 13.387 3.7545 13.6345 3.952L16.4845 2.76925L18.792 6.75L16.3133 8.623C16.3491 8.7845 16.3687 8.9355 16.372 9.076C16.3752 9.21633 16.3767 9.35767 16.3767 9.5C16.3767 9.63583 16.3735 9.774 16.367 9.9145C16.3607 10.0548 16.3377 10.2154 16.298 10.3963L18.7575 12.25L16.4498 16.25L13.6345 15.048C13.387 15.2455 13.1344 15.4202 12.8767 15.572C12.6191 15.724 12.355 15.8481 12.0845 15.9443L11.7037 19H7.08825ZM8.396 17.5H10.3615L10.721 14.8212C11.2313 14.6879 11.6977 14.4985 12.12 14.253C12.5425 14.0073 12.9499 13.6916 13.3422 13.3057L15.8267 14.35L16.8115 12.65L14.6422 11.0155C14.7256 10.7565 14.7823 10.5026 14.8123 10.2537C14.8424 10.0051 14.8575 9.75383 14.8575 9.5C14.8575 9.23967 14.8424 8.98842 14.8123 8.74625C14.7823 8.50392 14.7256 8.25642 14.6422 8.00375L16.8305 6.35L15.846 4.65L13.3325 5.7095C12.9978 5.35183 12.5969 5.03583 12.1297 4.7615C11.6624 4.48717 11.1897 4.29292 10.7115 4.17875L10.396 1.5H8.4115L8.0805 4.16925C7.57033 4.28975 7.09925 4.47433 6.66725 4.723C6.23508 4.97183 5.82283 5.29233 5.4305 5.6845L2.946 4.65L1.9615 6.35L4.121 7.9595C4.03767 8.19683 3.97933 8.44367 3.946 8.7C3.91267 8.95633 3.896 9.22617 3.896 9.5095C3.896 9.76983 3.91267 10.025 3.946 10.275C3.97933 10.525 4.0345 10.7718 4.1115 11.0155L1.9615 12.65L2.946 14.35L5.421 13.3C5.8005 13.6897 6.20625 14.0089 6.63825 14.2578C7.07042 14.5064 7.548 14.6974 8.071 14.8307L8.396 17.5ZM9.4075 12.5C10.2395 12.5 10.9475 12.208 11.5315 11.624C12.1155 11.04 12.4075 10.332 12.4075 9.5C12.4075 8.668 12.1155 7.96 11.5315 7.376C10.9475 6.792 10.2395 6.5 9.4075 6.5C8.56517 6.5 7.85458 6.792 7.27575 7.376C6.69692 7.96 6.4075 8.668 6.4075 9.5C6.4075 10.332 6.69692 11.04 7.27575 11.624C7.85458 12.208 8.56517 12.5 9.4075 12.5Z" />
-            </g>
-        </svg>
-    );
-}
-
-/**
- * 24x24 outer box holding a 16.596x16.596 glyph, inset 15.43%
- * on every side (3.701px). Fill #949494.
- */
-export function ExitToAppIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <g transform="translate(3.701 3.701)">
-                <path d="M1.705 16.596C1.23217 16.596 0.82975 16.4301 0.49775 16.0982C0.165917 15.7662 0 15.3632 0 14.8892V11.2787H1.399V14.8892C1.399 14.9662 1.431 15.0368 1.495 15.101C1.55917 15.165 1.62975 15.197 1.70675 15.197H14.8892C14.9662 15.197 15.0368 15.165 15.101 15.101C15.165 15.0368 15.197 14.9662 15.197 14.8892V1.70675C15.197 1.62975 15.165 1.55917 15.101 1.495C15.0368 1.431 14.9662 1.399 14.8892 1.399H1.70675C1.62975 1.399 1.55917 1.431 1.495 1.495C1.431 1.55917 1.399 1.62975 1.399 1.70675V5.3125H0V1.70675C0 1.23275 0.165917 0.829749 0.49775 0.497749C0.82975 0.165916 1.23217 0 1.705 0H14.891C15.3638 0 15.7662 0.165916 16.0982 0.497749C16.4301 0.829749 16.596 1.23217 16.596 1.705V14.891C16.596 15.3638 16.4301 15.7662 16.0982 16.0982C15.7662 16.4301 15.3638 16.596 14.891 16.596H1.705ZM6.9835 12.6382L5.998 11.6017L8.60475 8.995H0V7.596H8.60475L5.998 4.98925L6.9835 3.95775L11.3037 8.30575L6.9835 12.6382Z" />
-            </g>
-        </svg>
-    );
-}
-
-/** 21.5384x21.5384, fill #7C7C7C. */
-export function AddAPhotoIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 21.5384 21.5384"
-            width="21.5384"
-            height="21.5384"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M2.69208 18.8462C2.19849 18.8462 1.77595 18.6704 1.42445 18.3189C1.07296 17.9674 0.89721 17.5449 0.89721 17.0513V6.28211C0.89721 5.78852 1.07296 5.36598 1.42445 5.01449C1.77595 4.66299 2.19849 4.48724 2.69208 4.48724H5.51899L7.17925 2.69238H12.5639V4.48724H7.9645L6.32669 6.28211H2.69208V17.0513H17.051V8.97441H18.8459V17.0513C18.8459 17.5449 18.6701 17.9674 18.3186 18.3189C17.9672 18.6704 17.5446 18.8462 17.051 18.8462H2.69208ZM17.051 6.28211V4.48724H15.2562V2.69238H17.051V0.897509H18.8459V2.69238H20.6408V4.48724H18.8459V6.28211H17.051ZM9.87155 15.7052C10.9933 15.7052 11.9469 15.3125 12.7321 14.5273C13.5174 13.742 13.91 12.7885 13.91 11.6667C13.91 10.5449 13.5174 9.5914 12.7321 8.80614C11.9469 8.02089 10.9933 7.62826 9.87155 7.62826C8.74976 7.62826 7.79623 8.02089 7.01098 8.80614C6.22572 9.5914 5.8331 10.5449 5.8331 11.6667C5.8331 12.7885 6.22572 13.742 7.01098 14.5273C7.79623 15.3125 8.74976 15.7052 9.87155 15.7052ZM9.87155 13.9103C9.24335 13.9103 8.71236 13.6934 8.2786 13.2597C7.84484 12.8259 7.62796 12.2949 7.62796 11.6667C7.62796 11.0385 7.84484 10.5075 8.2786 10.0738C8.71236 9.64001 9.24335 9.42313 9.87155 9.42313C10.4998 9.42313 11.0307 9.64001 11.4645 10.0738C11.8983 10.5075 12.1151 11.0385 12.1151 11.6667C12.1151 12.2949 11.8983 12.8259 11.4645 13.2597C11.0307 13.6934 10.4998 13.9103 9.87155 13.9103Z" />
-        </svg>
-    );
-}
-
-/**
- * 24x24 outer box holding an 11.073x19.3075 chevron, inset
- * 1.44% left (0.346px) and 9.78% top/bottom (2.346px). Used by "Go back" on the
- * settings sub-nav.
- */
-export function ArrowBackIosIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <g transform="translate(0.346 2.346)">
-                <path d="M9.65375 19.3075L0 9.65375L9.65375 0L11.073 1.41925L2.83825 9.65375L11.073 17.8883L9.65375 19.3075Z" />
-            </g>
-        </svg>
-    );
-}
-
-/**
- * 24x24 outer box holding an 18x18 glyph, inset 12.5% (3px)
- * on every side. Lime when its sub-nav row is selected.
- */
-export function AccountCircleIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <g transform="translate(3 3)">
-                <path d="M3.19625 14.4845C4.04625 13.873 4.94817 13.3894 5.902 13.0337C6.85583 12.6779 7.8885 12.5 9 12.5C10.1115 12.5 11.1442 12.6779 12.098 13.0337C13.0518 13.3894 13.9538 13.873 14.8038 14.4845C15.4641 13.8012 15.9952 12.9942 16.397 12.0635C16.799 11.1327 17 10.1115 17 9C17 6.78333 16.2208 4.89583 14.6625 3.3375C13.1042 1.77917 11.2167 1 9 1C6.78333 1 4.89583 1.77917 3.3375 3.3375C1.77917 4.89583 1 6.78333 1 9C1 10.1115 1.201 11.1327 1.603 12.0635C2.00483 12.9942 2.53592 13.8012 3.19625 14.4845ZM9.0005 9.5C8.15783 9.5 7.44708 9.21075 6.86825 8.63225C6.28942 8.05375 6 7.34317 6 6.5005C6 5.65783 6.28925 4.94708 6.86775 4.36825C7.44625 3.78942 8.15683 3.5 8.9995 3.5C9.84217 3.5 10.5529 3.78925 11.1318 4.36775C11.7106 4.94625 12 5.65683 12 6.4995C12 7.34217 11.7107 8.05292 11.1322 8.63175C10.5537 9.21058 9.84317 9.5 9.0005 9.5ZM9 18C7.74483 18 6.56983 17.7663 5.475 17.299C4.38017 16.8317 3.42758 16.1929 2.61725 15.3828C1.80708 14.5724 1.16833 13.6198 0.701 12.525C0.233667 11.4302 0 10.2552 0 9C0 7.74483 0.233667 6.56983 0.701 5.475C1.16833 4.38017 1.80708 3.42758 2.61725 2.61725C3.42758 1.80708 4.38017 1.16833 5.475 0.700999C6.56983 0.233666 7.74483 0 9 0C10.2552 0 11.4302 0.233666 12.525 0.700999C13.6198 1.16833 14.5724 1.80708 15.3828 2.61725C16.1929 3.42758 16.8317 4.38017 17.299 5.475C17.7663 6.56983 18 7.74483 18 9C18 10.2552 17.7663 11.4302 17.299 12.525C16.8317 13.6198 16.1929 14.5724 15.3828 15.3828C14.5724 16.1929 13.6198 16.8317 12.525 17.299C11.4302 17.7663 10.2552 18 9 18ZM9 17C9.92183 17 10.8289 16.8388 11.7212 16.5163C12.6134 16.1939 13.3845 15.7526 14.0345 15.1923C13.3845 14.6704 12.6326 14.2579 11.7788 13.9548C10.9249 13.6516 9.99867 13.5 9 13.5C8.00133 13.5 7.07183 13.6484 6.2115 13.9453C5.35133 14.2421 4.60267 14.6578 3.9655 15.1923C4.6155 15.7526 5.38658 16.1939 6.27875 16.5163C7.17108 16.8388 8.07817 17 9 17ZM9 8.5C9.5615 8.5 10.0352 8.30708 10.4212 7.92125C10.8071 7.53525 11 7.0615 11 6.5C11 5.9385 10.8071 5.46475 10.4212 5.07875C10.0352 4.69292 9.5615 4.5 9 4.5C8.4385 4.5 7.96475 4.69292 7.57875 5.07875C7.19292 5.46475 7 5.9385 7 6.5C7 7.0615 7.19292 7.53525 7.57875 7.92125C7.96475 8.30708 8.4385 8.5 9 8.5Z" />
-            </g>
-        </svg>
-    );
-}
-
-/** 24x24; the glyph already fills its box in the design. */
-export function LockIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M6.30775 21.5C5.80908 21.5 5.38308 21.3234 5.02975 20.9703C4.67658 20.6169 4.5 20.1909 4.5 19.6923V10.3077C4.5 9.80908 4.67658 9.38308 5.02975 9.02975C5.38308 8.67658 5.80908 8.5 6.30775 8.5H7.5V6.5C7.5 5.25133 7.93783 4.18917 8.8135 3.3135C9.68917 2.43783 10.7513 2 12 2C13.2487 2 14.3108 2.43783 15.1865 3.3135C16.0622 4.18917 16.5 5.25133 16.5 6.5V8.5H17.6923C18.1909 8.5 18.6169 8.67658 18.9703 9.02975C19.3234 9.38308 19.5 9.80908 19.5 10.3077V19.6923C19.5 20.1909 19.3234 20.6169 18.9703 20.9703C18.6169 21.3234 18.1909 21.5 17.6923 21.5H6.30775ZM6.30775 20H17.6923C17.7821 20 17.8558 19.9712 17.9135 19.9135C17.9712 19.8558 18 19.7821 18 19.6923V10.3077C18 10.2179 17.9712 10.1442 17.9135 10.0865C17.8558 10.0288 17.7821 10 17.6923 10H6.30775C6.21792 10 6.14417 10.0288 6.0865 10.0865C6.02883 10.1442 6 10.2179 6 10.3077V19.6923C6 19.7821 6.02883 19.8558 6.0865 19.9135C6.14417 19.9712 6.21792 20 6.30775 20ZM12 16.75C12.4858 16.75 12.899 16.5798 13.2395 16.2395C13.5798 15.899 13.75 15.4858 13.75 15C13.75 14.5142 13.5798 14.101 13.2395 13.7605C12.899 13.4202 12.4858 13.25 12 13.25C11.5142 13.25 11.101 13.4202 10.7605 13.7605C10.4202 14.101 10.25 14.5142 10.25 15C10.25 15.4858 10.4202 15.899 10.7605 16.2395C11.101 16.5798 11.5142 16.75 12 16.75ZM9 8.5H15V6.5C15 5.66667 14.7083 4.95833 14.125 4.375C13.5417 3.79167 12.8333 3.5 12 3.5C11.1667 3.5 10.4583 3.79167 9.875 4.375C9.29167 4.95833 9 5.66667 9 6.5V8.5Z" />
-        </svg>
-    );
-}
-
-/**
- * The overflow trigger on a project row — 24x24.
- *
- * Ships from the design as a glyph clipped by an alpha mask whose only content is
- * the bounding-box rect; that mask is a no-op and is flattened away, as with the
- * other Material-derived icons above. The glyph already spans its box, so there is
- * no inner offset to preserve.
- */
-export function MoreVertIcon({ className }: IconProps) {
-    return (
-        <svg
-            viewBox="0 0 24 24"
-            width="24"
-            height="24"
-            fill="currentColor"
-            xmlns="http://www.w3.org/2000/svg"
-            aria-hidden="true"
-            focusable="false"
-            className={className}
-        >
-            <path d="M12 20C11.45 20 10.9792 19.8042 10.5875 19.4125C10.1958 19.0208 10 18.55 10 18C10 17.45 10.1958 16.9792 10.5875 16.5875C10.9792 16.1958 11.45 16 12 16C12.55 16 13.0208 16.1958 13.4125 16.5875C13.8042 16.9792 14 17.45 14 18C14 18.55 13.8042 19.0208 13.4125 19.4125C13.0208 19.8042 12.55 20 12 20ZM12 14C11.45 14 10.9792 13.8042 10.5875 13.4125C10.1958 13.0208 10 12.55 10 12C10 11.45 10.1958 10.9792 10.5875 10.5875C10.9792 10.1958 11.45 10 12 10C12.55 10 13.0208 10.1958 13.4125 10.5875C13.8042 10.9792 14 11.45 14 12C14 12.55 13.8042 13.0208 13.4125 13.4125C13.0208 13.8042 12.55 14 12 14ZM12 8C11.45 8 10.9792 7.80417 10.5875 7.4125C10.1958 7.02083 10 6.55 10 6C10 5.45 10.1958 4.97917 10.5875 4.5875C10.9792 4.19583 11.45 4 12 4C12.55 4 13.0208 4.19583 13.4125 4.5875C13.8042 4.97917 14 5.45 14 6C14 6.55 13.8042 7.02083 13.4125 7.4125C13.0208 7.80417 12.55 8 12 8Z" />
-        </svg>
-    );
-}
