+++ webapp/src/components/icons/menu-icon.tsx
@@ -1,24 +1,22 @@
 import { FigmaIconProps } from "./types";
 
 /*
- * Plus icon — a plus sign.
+ * Menu icon — three bars stacked one above another.
  */
 
-export function PlusIcon({ className }: FigmaIconProps) {
+export function MenuIcon({ className }: FigmaIconProps) {
     return (
         <svg
             viewBox="0 0 24 24"
             width="24"
             height="24"
+            fill="currentColor"
             xmlns="http://www.w3.org/2000/svg"
             aria-hidden="true"
             focusable="false"
             className={className}
         >
-            <path
-                fill="currentColor"
-                d="M19 12.998h-6v6h-2v-6H5v-2h6v-6h2v6h6z"
-            />
+            <path d="M3 18V16H21V18H3ZM3 13V11H21V13H3ZM3 8V6H21V8H3Z" />
         </svg>
     );
 }
