+++ webapp/src/components/icons/settings-icon.tsx
@@ -1,0 +1,24 @@
+import { FigmaIconProps } from "./types";
+
+/*
+ * Settings icon — a cogwheel.
+ */
+
+export function SettingsIcon({ className }: FigmaIconProps) {
+    return (
+        <svg
+            viewBox="0 0 24 24"
+            width="24"
+            height="24"
+            fill="currentColor"
+            xmlns="http://www.w3.org/2000/svg"
+            aria-hidden="true"
+            focusable="false"
+            className={className}
+        >
+            <g transform="translate(2.604 2.5)">
+                <path d="M7.08825 19L6.7075 15.9538C6.43967 15.8641 6.165 15.7385 5.8835 15.577C5.60217 15.4153 5.35058 15.2422 5.12875 15.0577L2.3075 16.25L0 12.25L2.44025 10.4058C2.41725 10.2571 2.40092 10.1078 2.39125 9.95775C2.38158 9.80775 2.37675 9.65833 2.37675 9.5095C2.37675 9.36733 2.38158 9.22283 2.39125 9.076C2.40092 8.92917 2.41725 8.76858 2.44025 8.59425L0 6.75L2.3075 2.76925L5.119 3.952C5.36017 3.761 5.61758 3.58633 5.89125 3.428C6.16492 3.26967 6.43383 3.14242 6.698 3.04625L7.08825 0H11.7037L12.0845 3.05575C12.3845 3.16475 12.6559 3.292 12.8987 3.4375C13.1417 3.583 13.387 3.7545 13.6345 3.952L16.4845 2.76925L18.792 6.75L16.3133 8.623C16.3491 8.7845 16.3687 8.9355 16.372 9.076C16.3752 9.21633 16.3767 9.35767 16.3767 9.5C16.3767 9.63583 16.3735 9.774 16.367 9.9145C16.3607 10.0548 16.3377 10.2154 16.298 10.3963L18.7575 12.25L16.4498 16.25L13.6345 15.048C13.387 15.2455 13.1344 15.4202 12.8767 15.572C12.6191 15.724 12.355 15.8481 12.0845 15.9443L11.7037 19H7.08825ZM8.396 17.5H10.3615L10.721 14.8212C11.2313 14.6879 11.6977 14.4985 12.12 14.253C12.5425 14.0073 12.9499 13.6916 13.3422 13.3057L15.8267 14.35L16.8115 12.65L14.6422 11.0155C14.7256 10.7565 14.7823 10.5026 14.8123 10.2537C14.8424 10.0051 14.8575 9.75383 14.8575 9.5C14.8575 9.23967 14.8424 8.98842 14.8123 8.74625C14.7823 8.50392 14.7256 8.25642 14.6422 8.00375L16.8305 6.35L15.846 4.65L13.3325 5.7095C12.9978 5.35183 12.5969 5.03583 12.1297 4.7615C11.6624 4.48717 11.1897 4.29292 10.7115 4.17875L10.396 1.5H8.4115L8.0805 4.16925C7.57033 4.28975 7.09925 4.47433 6.66725 4.723C6.23508 4.97183 5.82283 5.29233 5.4305 5.6845L2.946 4.65L1.9615 6.35L4.121 7.9595C4.03767 8.19683 3.97933 8.44367 3.946 8.7C3.91267 8.95633 3.896 9.22617 3.896 9.5095C3.896 9.76983 3.91267 10.025 3.946 10.275C3.97933 10.525 4.0345 10.7718 4.1115 11.0155L1.9615 12.65L2.946 14.35L5.421 13.3C5.8005 13.6897 6.20625 14.0089 6.63825 14.2578C7.07042 14.5064 7.548 14.6974 8.071 14.8307L8.396 17.5ZM9.4075 12.5C10.2395 12.5 10.9475 12.208 11.5315 11.624C12.1155 11.04 12.4075 10.332 12.4075 9.5C12.4075 8.668 12.1155 7.96 11.5315 7.376C10.9475 6.792 10.2395 6.5 9.4075 6.5C8.56517 6.5 7.85458 6.792 7.27575 7.376C6.69692 7.96 6.4075 8.668 6.4075 9.5C6.4075 10.332 6.69692 11.04 7.27575 11.624C7.85458 12.208 8.56517 12.5 9.4075 12.5Z" />
+            </g>
+        </svg>
+    );
+}
